REF: useWatchConnectivity
