--- components/HandOffComponentListener.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const HandOffComponentListener: React.FC = () => {
-  return null;
-};
-
-export default HandOffComponentListener;
