+++ hooks/useHandOffListener.ts
@@ -1,0 +1,4 @@
+
+const useHandOffListener = () => {};
+
+export default useHandOffListener;
