+++ hooks/useWatchConnectivity.ts
@@ -1,0 +1,3 @@
+function useWatchConnectivity() {}
+
+export default useWatchConnectivity;
